Extract shared KerbalsTestHelper from 4 duplicate test copies
